增加表格通用查询功能

+++ platform-application/src/main/java/com/pisces/platform/application/query/GridResponse.java
@@ -14,6 +14,6 @@ import java.util.List;
  */
 @Data
 public class GridResponse {
-    private Integer totalRow = 0;
+    private Long totalRow = 0L;
     private List<? extends BeanObject> beans = new ArrayList<>();
 }
